instaleads/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
aiosqlite>=0.20.0
pydantic>=2.7.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
curl_cffi>=0.7.0
git+https://github.com/subzeroid/instagrapi.git
cryptography>=44.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
